@@ -1,7 +1,8 @@
 {
   "private": true,
+  "type": "module",
   "name": "illustrative-code",
-  "version": "1.2.0",
+  "version": "2.0.0",
   "description": "Illustrations made with code",
   "repository": {
     "type": "git",
@@ -12,9 +13,13 @@
     "build": "NODE_ENV=production npx gulp build",
     "develop": "NODE_ENV=development npx gulp develop",
     "dev": "npm run develop",
+    "lint": "npx eslint --cache --ignore-path .gitignore ./components/**/*.js ./*.js",
     "serve": "npx gulp serve",
     "start": "npm run develop"
   },
+  "pre-commit": [
+    "lint"
+  ],
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/joeleisner/illustrative-code/issues"
@@ -23,29 +28,35 @@
   "devDependencies": {
     "@babel/core": "^7.12.3",
     "@babel/preset-env": "^7.12.1",
-    "@babel/register": "^7.12.1",
-    "autoprefixer": "^9.8.6",
+    "autoprefixer": "^10.3.1",
+    "babel-eslint": "^10.1.0",
     "babel-loader": "^8.1.0",
     "browser-sync": "^2.26.13",
-    "cssnano": "^4.1.10",
-    "eslint": "^7.13.0",
+    "cssnano": "^5.0.7",
+    "eslint": "^7.32.0",
     "fibers": "^5.0.0",
     "gulp": "^4.0.2",
     "gulp-image-resize": "^0.13.1",
-    "gulp-imagemin": "^7.1.0",
+    "gulp-imagemin": "^8.0.0",
     "gulp-postcss": "^9.0.0",
-    "gulp-pug": "^4.0.1",
+    "gulp-pug": "^5.0.0",
     "gulp-rename": "^2.0.0",
-    "gulp-sass": "^4.1.0",
+    "gulp-sass": "^5.0.0",
     "jstransformer-markdown-it": "^2.1.0",
+    "pre-commit": "^1.2.2",
     "sass": "^1.29.0",
     "terser-webpack-plugin": "^5.0.3",
-    "vinyl-named": "^1.1.0",
+    "vinyl-named-with-path": "^1.0.0",
     "webpack": "^5.4.0",
     "webpack-stream": "^6.1.0"
   },
-  "browserslist": "defaults and last 2 versions, not Explorer <= 11, not ExplorerMobile 11",
+  "browserslist": [
+    "defaults and last 2 versions",
+    "not Explorer <= 11",
+    "not ExplorerMobile 11",
+    "not OperaMini all"
+  ],
   "dependencies": {
-    "js-cookie": "^2.2.1"
+    "js-cookie": "^3.0.0"
   }
 }
